Add container to settings community page
