+++ routes/admin/authenticated/sidebar/community-settings/mailchimp/index.js
@@ -2,7 +2,7 @@
 if (typeof require.ensure !== 'function') require.ensure = (d, c) => c(require)
 
 export default store => ({
-  path: 'community/recipient',
+  path: 'mailchimp',
   getComponent (nextState, callback) {
     require.ensure([], function (require) {
       callback(null, require('./page.connected').default)
