+++ routes/admin/authenticated/sidebar/community-settings/report/page.connected.js
@@ -1,0 +1,17 @@
+import { connect } from 'react-redux'
+
+import * as CommunityActions from '~community/action-creators'
+import * as CommunitySelectors from '~community/selectors'
+
+import Page from './page'
+
+const mapStateToProps = state => ({
+  community: CommunitySelectors.getCurrent(state)
+})
+
+const mapDispatchToProps = {
+  submit: CommunityActions.asyncEdit,
+  ...CommunityActions
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Page)
