+++ test/routes/unit/logged-in/sidebar/community-settings/info/page.spec.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { shallow } from 'enzyme'
 
 import * as mock from '~client/utils/mock'
-import Page from '~routes/admin/authenticated/sidebar/community-settings-info/page'
+import Page from '~routes/admin/authenticated/sidebar/community-settings/info/page'
 
-describe('routes/admin/authenticated/sidebar/community-settings-info/page', () => {
+describe('routes/admin/authenticated/sidebar/community-settings/info/page', () => {
   const props = {
     fields: {
       image: {},
