+++ test/routes/unit/logged-in/sidebar/community-settings/mailchimp/page.spec.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { shallow } from 'enzyme'
 
 import * as mock from '~client/utils/mock'
-import Page from '~routes/admin/authenticated/sidebar/community-settings-mailchimp/page'
+import Page from '~routes/admin/authenticated/sidebar/community-settings/mailchimp/page'
 
-describe('routes/admin/authenticated/sidebar/community-settings-mailchimp/page', () => {
+describe('routes/admin/authenticated/sidebar/community-settings/mailchimp/page', () => {
   const props = {
     fields: {
       mailchimp_api_key: {},
